Self test adapted to latest changes in package code (2)
and: TSM_0150 and TSM_0151 commented out
