+++ test/pytest/pytestfiles/test_08_VERSION_CONTROL_BADCASE.py
@@ -14,11 +14,11 @@
 #  limitations under the License.
 # --------------------------------------------------------------------------------------------------------------
 #
-# test_09_VERSION_CONTROL_BADCASE.py
+# test_08_VERSION_CONTROL_BADCASE.py
 #
 # XC-CT/ECA3-Queckenstedt
 #
-# 23.05.2023 - 15:45:58
+# 31.05.2023 - 15:27:30
 #
 # --------------------------------------------------------------------------------------------------------------
