+++ test/pytest/pytestfiles/test_09_ROBOT_CODE_BADCASE.py
@@ -14,11 +14,11 @@
 #  limitations under the License.
 # --------------------------------------------------------------------------------------------------------------
 #
-# test_10_ROBOT_CODE_BADCASE.py
+# test_09_ROBOT_CODE_BADCASE.py
 #
 # XC-CT/ECA3-Queckenstedt
 #
-# 24.05.2023 - 20:42:16
+# 31.05.2023 - 15:27:30
 #
 # --------------------------------------------------------------------------------------------------------------
