+++ test/referencelogfiles/TSM_0001-(VARIANT_HANDLING)-[GOODCASE]/TSM_0001-(VARIANT_HANDLING)-[GOODCASE].log
@@ -1,44 +1,44 @@
 ==============================================================================
-20230503 13:54:11.762 - INFO - + START SUITE: Tsm-Testfile-01 [ ]
+20230531 14:32:55.852 - INFO - + START SUITE: Tsm-Testfile-01 [ ]
 ==============================================================================
-20230503 13:54:11.774 - INFO - +- START SETUP: tm.Testsuite Setup [ ]
-20230503 13:54:11.804 - INFO - Set suite metadata 'project' to value 'RobotFramework Testsuites'.
-20230503 13:54:11.804 - INFO - Set suite metadata 'machine' to value 'HI7-C-0005Y'.
-20230503 13:54:11.806 - INFO - Set suite metadata 'tester' to value 'Queckenstedt Holger (XC-CT/ECA3)'.
-20230503 13:54:11.806 - INFO - Set suite metadata 'testtool' to value 'Robot Framework 4.1.3 (Python 3.9.16 on win32)'.
-20230503 13:54:11.806 - INFO - Set suite metadata 'version' to value '0.7.0'.
-20230503 13:54:11.806 - INFO - Set suite metadata 'version_sw' to value ''.
-20230503 13:54:11.806 - INFO - Set suite metadata 'version_hw' to value ''.
-20230503 13:54:11.806 - INFO - Set suite metadata 'version_test' to value ''.
-20230503 13:54:11.807 - INFO - ${CONFIG} = {'Project': 'RobotFramework Testsuites', 'WelcomeString': 'Hello... RobotFramework AIO is running now!', 'Maximum_version': '1.0.0', 'Minimum_version': '0.6.0', 'version': {'majorversion': '0', 'minor...
-20230503 13:54:11.807 - WARN - Running with configuration level 4! 
-Selected configuration file 'D:\RobotFramework\python39\lib\site-packages\RobotFramework_TestsuitesManagement\Config\robot_config.json'
-20230503 13:54:11.807 - INFO - Robot Framework AIO version check passed!
-20230503 13:54:11.807 - INFO - Suite Path: 'C:\RobotTest\projects\ROBFW_AIO\robotframework-testsuitesmanagement\test\testfiles\'
-20230503 13:54:11.807 - INFO - CfgFile Path: 'D:\RobotFramework\python39\lib\site-packages\RobotFramework_TestsuitesManagement\Config\robot_config.json'
-20230503 13:54:11.808 - INFO - Number of test suites: 1
-20230503 13:54:11.808 - INFO - Total number of testcases: 1
-20230503 13:54:11.808 - INFO - +- END SETUP: tm.Testsuite Setup (34)
+20230531 14:32:55.856 - INFO - +- START SETUP: tm.Testsuite Setup [ ]
+20230531 14:32:55.862 - INFO - Set suite metadata 'project' to value 'RobotFramework Testsuites'.
+20230531 14:32:55.878 - INFO - Set suite metadata 'machine' to value 'HI7-C-0005Y'.
+20230531 14:32:55.880 - INFO - Set suite metadata 'tester' to value 'Queckenstedt Holger (XC-CT/ECA3)'.
+20230531 14:32:55.881 - INFO - Set suite metadata 'testtool' to value 'Robot Framework 6.0.2 (Python 3.9.16 on win32)'.
+20230531 14:32:55.881 - INFO - Set suite metadata 'bundle_version' to value '0.8.0.0'.
+20230531 14:32:55.881 - INFO - Set suite metadata 'version_sw' to value ''.
+20230531 14:32:55.881 - INFO - Set suite metadata 'version_hw' to value ''.
+20230531 14:32:55.881 - INFO - Set suite metadata 'version_test' to value ''.
+20230531 14:32:55.882 - INFO - ${CONFIG} = {'Project': 'RobotFramework Testsuites', 'WelcomeString': 'Hello... RobotFramework AIO is running now!', 'Maximum_version': '1.0.0', 'Minimum_version': '0.6.0', 'version': {'majorversion': '0', 'minor...
+20230531 14:32:55.882 - WARN - Running with configuration level 4! 
+Selected configuration file 'D:/RobotFramework/python39/lib/site-packages/RobotFramework_TestsuitesManagement/Config/robot_config.jsonp'
+20230531 14:32:55.883 - INFO - RobotFramework AIO version check passed!
+20230531 14:32:55.883 - INFO - Suite Path: 'C:\RobotTest\projects\ROBFW_AIO\robotframework-testsuitesmanagement\test\testfiles\'
+20230531 14:32:55.883 - INFO - CfgFile Path: 'D:/RobotFramework/python39/lib/site-packages/RobotFramework_TestsuitesManagement/Config/robot_config.jsonp'
+20230531 14:32:55.883 - INFO - Number of test suites: 1
+20230531 14:32:55.883 - INFO - Total number of testcases: 1
+20230531 14:32:55.883 - INFO - +- END SETUP: tm.Testsuite Setup (27)
 ------------------------------------------------------------------------------
-20230503 13:54:11.808 - INFO - +- START TEST: Test Case tsm-testfile-01 [ ]
+20230531 14:32:55.883 - INFO - +- START TEST: Test Case tsm-testfile-01 [ ]
 ------------------------------------------------------------------------------
-20230503 13:54:11.809 - INFO - +-- START SETUP: tm.Testcase Setup [ ]
-20230503 13:54:11.809 - INFO - Test Count: 1
-20230503 13:54:11.810 - INFO - +-- END SETUP: tm.Testcase Setup (0)
+20230531 14:32:55.884 - INFO - +-- START SETUP: tm.Testcase Setup [ ]
+20230531 14:32:55.884 - INFO - Test Count: 1
+20230531 14:32:55.884 - INFO - +-- END SETUP: tm.Testcase Setup (0)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-20230503 13:54:11.810 - INFO - +-- START KEYWORD: BuiltIn.Log [ tsm-testfile-01 | console=yes ]
-20230503 13:54:11.810 - INFO - tsm-testfile-01
-20230503 13:54:11.811 - INFO - +-- END KEYWORD: BuiltIn.Log (1)
+20230531 14:32:55.884 - INFO - +-- START KEYWORD: BuiltIn.Log [ tsm-testfile-01 | console=yes ]
+20230531 14:32:55.885 - INFO - tsm-testfile-01
+20230531 14:32:55.885 - INFO - +-- END KEYWORD: BuiltIn.Log (1)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-20230503 13:54:11.811 - INFO - +-- START TEARDOWN: tm.Testcase Teardown [ ]
-20230503 13:54:11.811 - INFO - testcase_teardown: Will be implemented later
-20230503 13:54:11.811 - INFO - +-- END TEARDOWN: tm.Testcase Teardown (0)
+20230531 14:32:55.886 - INFO - +-- START TEARDOWN: tm.Testcase Teardown [ ]
+20230531 14:32:55.886 - INFO - testcase_teardown: Will be implemented later
+20230531 14:32:55.886 - INFO - +-- END TEARDOWN: tm.Testcase Teardown (0)
 ------------------------------------------------------------------------------
-20230503 13:54:11.812 - INFO - +- END TEST: Test Case tsm-testfile-01 (3)
+20230531 14:32:55.887 - INFO - +- END TEST: Test Case tsm-testfile-01 (3)
 ------------------------------------------------------------------------------
-20230503 13:54:11.813 - INFO - +- START TEARDOWN: tm.Testsuite Teardown [ ]
-20230503 13:54:11.813 - INFO - testsuite_teardown: Will be implemented later
-20230503 13:54:11.813 - INFO - +- END TEARDOWN: tm.Testsuite Teardown (0)
+20230531 14:32:55.888 - INFO - +- START TEARDOWN: tm.Testsuite Teardown [ ]
+20230531 14:32:55.888 - INFO - testsuite_teardown: Will be implemented later
+20230531 14:32:55.888 - INFO - +- END TEARDOWN: tm.Testsuite Teardown (0)
 ==============================================================================
-20230503 13:54:11.815 - INFO - + END SUITE: Tsm-Testfile-01 (145)
+20230531 14:32:55.890 - INFO - + END SUITE: Tsm-Testfile-01 (131)
 ==============================================================================
