--- test/referencelogfiles/TSM_0150-(LOCAL_CONFIG)-[BADCASE]/TSM_0150-(LOCAL_CONFIG)-[BADCASE].log
@@ -1,19 +1,0 @@
-==============================================================================
-20230531 12:48:44.538 - INFO - + START SUITE: Tsm-Testfile-01 [ ]
-==============================================================================
-20230531 12:48:44.552 - INFO - +- START SETUP: tm.Testsuite Setup [ ]
-20230531 12:48:44.552 - ERROR - Did not find configuration file: 'D:/RobotFramework/python39/lib/site-packages/RobotFramework_TestsuitesManagement/Config/robot_config.json'!
-20230531 12:48:44.552 - UNKNOWN - The configuration file is not found!
-20230531 12:48:44.552 - DEBUG - The configuration file is not found!
-20230531 12:48:44.552 - INFO - +- END SETUP: tm.Testsuite Setup (0)
-------------------------------------------------------------------------------
-20230531 12:48:44.552 - INFO - +- START TEST: Test Case tsm-testfile-01 [ ]
-------------------------------------------------------------------------------
-20230531 12:48:44.552 - INFO - +- END TEST: Test Case tsm-testfile-01 (0)
-------------------------------------------------------------------------------
-20230531 12:48:44.552 - INFO - +- START TEARDOWN: tm.Testsuite Teardown [ ]
-20230531 12:48:44.552 - INFO - testsuite_teardown: Will be implemented later
-20230531 12:48:44.552 - INFO - +- END TEARDOWN: tm.Testsuite Teardown (0)
-==============================================================================
-20230531 12:48:44.552 - INFO - + END SUITE: Tsm-Testfile-01 (100)
-==============================================================================
