--- test/referencelogfiles/TSM_0150-(LOCAL_CONFIG)-[BADCASE]/TSM_0150-(LOCAL_CONFIG)-[BADCASE].xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<robot generator="Robot 6.0.2 (Python 3.9.16 on win32)" generated="20230531 12:48:44.438" rpa="false" schemaversion="3">
-<suite id="s1" name="Tsm-Testfile-01" source="C:\RobotTest\projects\ROBFW_AIO\robotframework-testsuitesmanagement\test\testfiles\tsm-testfile-01.robot">
-<kw name="Testsuite Setup" library="tm" type="SETUP">
-<doc>This testsuite_setup defines the ``Testsuite Setup`` which is used to loads the RobotFramework AIO configuration,
-checks the version of RobotFramework AIO, and logs out the basic information of the robot run.</doc>
-<msg timestamp="20230531 12:48:44.552" level="ERROR">Did not find configuration file: 'D:/RobotFramework/python39/lib/site-packages/RobotFramework_TestsuitesManagement/Config/robot_config.json'!
-</msg>
-<msg timestamp="20230531 12:48:44.552" level="UNKNOWN">The configuration file is not found!</msg>
-<status status="UNKNOWN" starttime="20230531 12:48:44.552" endtime="20230531 12:48:44.552"/>
-</kw>
-<test id="s1-t1" name="Test Case tsm-testfile-01" line="31">
-<doc>tsm-testfile-01</doc>
-<status status="UNKNOWN" starttime="20230531 12:48:44.552" endtime="20230531 12:48:44.552">Parent suite setup unknown:
-The configuration file is not found!</status>
-</test>
-<kw name="Testsuite Teardown" library="tm" type="TEARDOWN">
-<doc>This testsuite_teardown defines the ``Testsuite Teardown`` keyword, currently this keyword does nothing,
-it's defined here for future requirements.</doc>
-<msg timestamp="20230531 12:48:44.552" level="INFO">testsuite_teardown: Will be implemented later</msg>
-<status status="PASS" starttime="20230531 12:48:44.552" endtime="20230531 12:48:44.552"/>
-</kw>
-<status status="UNKNOWN" starttime="20230531 12:48:44.452" endtime="20230531 12:48:44.552">Suite setup unknown:
-The configuration file is not found!</status>
-</suite>
-<statistics>
-<total>
-<stat pass="0" fail="0" skip="0" unknown="1">All Tests</stat>
-</total>
-<tag>
-</tag>
-<suite>
-<stat pass="0" fail="0" skip="0" unknown="1" id="s1" name="Tsm-Testfile-01">Tsm-Testfile-01</stat>
-</suite>
-</statistics>
-<errors>
-<msg timestamp="20230531 12:48:44.552" level="ERROR">Did not find configuration file: 'D:/RobotFramework/python39/lib/site-packages/RobotFramework_TestsuitesManagement/Config/robot_config.json'!
-</msg>
-</errors>
-</robot>
