--- test/referencelogfiles/TSM_0151-(LOCAL_CONFIG)-[BADCASE]/TSM_0151-(LOCAL_CONFIG)-[BADCASE].log
@@ -1,50 +1,0 @@
-==============================================================================
-20230531 12:48:45.052 - INFO - + START SUITE: Tsm-Testfile-02 [ ]
-==============================================================================
-20230531 12:48:45.068 - INFO - +- START SETUP: tm.Testsuite Setup [ ./config/tsm-test_variants.jsonp ]
-20230531 12:48:45.099 - FAIL - RecursionError: maximum recursion depth exceeded while calling a Python object
-20230531 12:48:45.099 - DEBUG - Traceback (most recent call last):
-  File "D:\RobotFramework\python39\lib\site-packages\RobotFramework_TestsuitesManagement\__init__.py", line 65, in run_keyword
-    return DynamicCore.run_keyword(self, name, args, kwargs)
-  File "D:\RobotFramework\python39\lib\site-packages\robotlibcore.py", line 122, in run_keyword
-    return self.keywords[name](*args, **(kwargs or {}))
-  File "D:\RobotFramework\python39\lib\site-packages\RobotFramework_TestsuitesManagement\Keywords\CSetup.py", line 77, in testsuite_setup
-    RobotFramework_TestsuitesManagement.CTestsuitesCfg.oConfig.loadCfg(RobotFramework_TestsuitesManagement.CTestsuitesCfg.oConfig)
-  File "D:\RobotFramework\python39\lib\site-packages\RobotFramework_TestsuitesManagement\Config\CConfig.py", line 380, in loadCfg
-    __loadLocalConfig(oLocalConfig)
-  File "D:\RobotFramework\python39\lib\site-packages\RobotFramework_TestsuitesManagement\Config\CConfig.py", line 378, in __loadLocalConfig
-    __loadLocalConfig(oLocalConfig)
-  File "D:\RobotFramework\python39\lib\site-packages\RobotFramework_TestsuitesManagement\Config\CConfig.py", line 378, in __loadLocalConfig
-    __loadLocalConfig(oLocalConfig)
-  File "D:\RobotFramework\python39\lib\site-packages\RobotFramework_TestsuitesManagement\Config\CConfig.py", line 378, in __loadLocalConfig
-    __loadLocalConfig(oLocalConfig)
-  [Previous line repeated 964 more times]
-  File "D:\RobotFramework\python39\lib\site-packages\RobotFramework_TestsuitesManagement\Config\CConfig.py", line 373, in __loadLocalConfig
-    tmpDict = copy.deepcopy(oLocalConfig)
-  File "D:\RobotFramework\python39\lib\copy.py", line 146, in deepcopy
-    y = copier(x, memo)
-  File "D:\RobotFramework\python39\lib\copy.py", line 230, in _deepcopy_dict
-    y[deepcopy(key, memo)] = deepcopy(value, memo)
-  File "D:\RobotFramework\python39\lib\copy.py", line 146, in deepcopy
-    y = copier(x, memo)
-  File "D:\RobotFramework\python39\lib\copy.py", line 230, in _deepcopy_dict
-    y[deepcopy(key, memo)] = deepcopy(value, memo)
-  File "D:\RobotFramework\python39\lib\copy.py", line 146, in deepcopy
-    y = copier(x, memo)
-  File "D:\RobotFramework\python39\lib\copy.py", line 230, in _deepcopy_dict
-    y[deepcopy(key, memo)] = deepcopy(value, memo)
-  File "D:\RobotFramework\python39\lib\copy.py", line 137, in deepcopy
-    d = id(x)
-RecursionError: maximum recursion depth exceeded while calling a Python object
-20230531 12:48:45.099 - INFO - +- END SETUP: tm.Testsuite Setup (31)
-------------------------------------------------------------------------------
-20230531 12:48:45.099 - INFO - +- START TEST: Test Case tsm-testfile-02 [ ]
-------------------------------------------------------------------------------
-20230531 12:48:45.099 - INFO - +- END TEST: Test Case tsm-testfile-02 (0)
-------------------------------------------------------------------------------
-20230531 12:48:45.099 - INFO - +- START TEARDOWN: tm.Testsuite Teardown [ ]
-20230531 12:48:45.099 - INFO - testsuite_teardown: Will be implemented later
-20230531 12:48:45.099 - INFO - +- END TEARDOWN: tm.Testsuite Teardown (0)
-==============================================================================
-20230531 12:48:45.099 - INFO - + END SUITE: Tsm-Testfile-02 (160)
-==============================================================================
