--- test/referencelogfiles/TSM_0151-(LOCAL_CONFIG)-[BADCASE]/TSM_0151-(LOCAL_CONFIG)-[BADCASE].xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<robot generator="Robot 6.0.2 (Python 3.9.16 on win32)" generated="20230531 12:48:44.939" rpa="false" schemaversion="3">
-<suite id="s1" name="Tsm-Testfile-02" source="C:\RobotTest\projects\ROBFW_AIO\robotframework-testsuitesmanagement\test\testfiles\tsm-testfile-02.robot">
-<kw name="Testsuite Setup" library="tm" type="SETUP">
-<arg>./config/tsm-test_variants.jsonp</arg>
-<doc>This testsuite_setup defines the ``Testsuite Setup`` which is used to loads the RobotFramework AIO configuration,
-checks the version of RobotFramework AIO, and logs out the basic information of the robot run.</doc>
-<msg timestamp="20230531 12:48:45.099" level="FAIL">RecursionError: maximum recursion depth exceeded while calling a Python object</msg>
-<status status="FAIL" starttime="20230531 12:48:45.068" endtime="20230531 12:48:45.099"/>
-</kw>
-<test id="s1-t1" name="Test Case tsm-testfile-02" line="32">
-<doc>tsm-testfile-02</doc>
-<status status="FAIL" starttime="20230531 12:48:45.099" endtime="20230531 12:48:45.099">Parent suite setup failed:
-RecursionError: maximum recursion depth exceeded while calling a Python object</status>
-</test>
-<kw name="Testsuite Teardown" library="tm" type="TEARDOWN">
-<doc>This testsuite_teardown defines the ``Testsuite Teardown`` keyword, currently this keyword does nothing,
-it's defined here for future requirements.</doc>
-<msg timestamp="20230531 12:48:45.099" level="INFO">testsuite_teardown: Will be implemented later</msg>
-<status status="PASS" starttime="20230531 12:48:45.099" endtime="20230531 12:48:45.099"/>
-</kw>
-<status status="FAIL" starttime="20230531 12:48:44.939" endtime="20230531 12:48:45.099">Suite setup failed:
-RecursionError: maximum recursion depth exceeded while calling a Python object</status>
-</suite>
-<statistics>
-<total>
-<stat pass="0" fail="1" skip="0" unknown="0">All Tests</stat>
-</total>
-<tag>
-</tag>
-<suite>
-<stat pass="0" fail="1" skip="0" unknown="0" id="s1" name="Tsm-Testfile-02">Tsm-Testfile-02</stat>
-</suite>
-</statistics>
-<errors>
-</errors>
-</robot>
